+++ src/clj/withings_client/routes/services.clj
@@ -21,17 +21,17 @@
   ["/token/:id/refresh"
    {:post (fn [{{:keys [id]} :path-params}]
             (log/info "/token/:n/refresh" id)
-            (if (tokens/refresh-and-restore-one! id)
+            (if (tokens/refresh-and-restore-id! id)
               (response/ok "success")
               (response/bad-request "fail")))}]
-  
+
   ;; post?
-  ["/token/refresh-all"
+  ["/tokens/refresh-all"
     {:post (fn [_]
-             (log/info "/token/refresh-all")
+             (log/info "/tokens/refresh-all")
              (try
-               (tokens/refresh-all)
-               (response/ok "refresh all")
+               (tokens/refresh-all!)
+               (response/ok "refreshed all")
                (catch Exception e (error e))))}]
 
   ;; users